Rename AzureDataLakeStorage to ADLS
