+++ airflow/providers/microsoft/azure/example_dags/example_adls_delete.py
@@ -18,7 +18,7 @@
 import os
 
 from airflow import models
-from airflow.providers.microsoft.azure.operators.adls_delete import AzureDataLakeStorageDeleteOperator
+from airflow.providers.microsoft.azure.operators.adls import ADLSDeleteOperator
 from airflow.providers.microsoft.azure.transfers.local_to_adls import LocalFilesystemToADLSOperator
 from airflow.utils.dates import days_ago
 
@@ -39,9 +39,7 @@ with models.DAG(
         remote_path=REMOTE_FILE_PATH,
     )
     # [START howto_operator_adls_delete]
-    remove_file = AzureDataLakeStorageDeleteOperator(
-        task_id="delete_task", path=REMOTE_FILE_PATH, recursive=True
-    )
+    remove_file = ADLSDeleteOperator(task_id="delete_task", path=REMOTE_FILE_PATH, recursive=True)
     # [END howto_operator_adls_delete]
 
     upload_file >> remove_file
