+++ airflow/providers/microsoft/azure/operators/adls_delete.py
@@ -14,50 +14,25 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module is deprecated. Please use :mod:`airflow.providers.microsoft.azure.operators.adls`."""
 
-from typing import Any, Sequence
+import warnings
 
-from airflow.models import BaseOperator
-from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
+from airflow.providers.microsoft.azure.operators.adls import ADLSDeleteOperator
 
 
-class AzureDataLakeStorageDeleteOperator(BaseOperator):
+class AzureDataLakeStorageDeleteOperator(ADLSDeleteOperator):
     """
-    Delete files in the specified path.
-
-        .. seealso::
-            For more information on how to use this operator, take a look at the guide:
-            :ref:`howto/operator:AzureDataLakeStorageDeleteOperator`
-
-    :param path: A directory or file to remove
-    :type path: str
-    :param recursive: Whether to loop into directories in the location and remove the files
-    :type recursive: bool
-    :param ignore_not_found: Whether to raise error if file to delete is not found
-    :type ignore_not_found: bool
-    :param azure_data_lake_conn_id: Reference to the :ref:`Azure Data Lake connection<howto/connection:adl>`.
-    :type azure_data_lake_conn_id: str
+    This class is deprecated.
+    Please use `airflow.providers.microsoft.azure.operators.adls.ADLSDeleteOperator`.
     """
 
-    template_fields: Sequence[str] = ('path',)
-    ui_color = '#901dd2'
-
-    def __init__(
-        self,
-        *,
-        path: str,
-        recursive: bool = False,
-        ignore_not_found: bool = True,
-        azure_data_lake_conn_id: str = 'azure_data_lake_default',
-        **kwargs,
-    ) -> None:
-        super().__init__(**kwargs)
-        self.path = path
-        self.recursive = recursive
-        self.ignore_not_found = ignore_not_found
-        self.azure_data_lake_conn_id = azure_data_lake_conn_id
-
-    def execute(self, context: dict) -> Any:
-        hook = AzureDataLakeHook(azure_data_lake_conn_id=self.azure_data_lake_conn_id)
-
-        return hook.remove(path=self.path, recursive=self.recursive, ignore_not_found=self.ignore_not_found)
+    def __init__(self, *args, **kwargs):
+        warnings.warn(
+            """This class is deprecated.
+            Please use
+            `airflow.providers.microsoft.azure.operators.adls.ADLSDeleteOperator`.""",
+            DeprecationWarning,
+            stacklevel=3,
+        )
+        super().__init__(*args, **kwargs)
