+++ tests/providers/google/cloud/transfers/test_adls_to_gcs.py
@@ -57,7 +57,7 @@ class TestAdlsToGoogleCloudStorageOperator(unittest.TestCase):
         assert operator.azure_data_lake_conn_id == AZURE_CONN_ID
 
     @mock.patch('airflow.providers.google.cloud.transfers.adls_to_gcs.AzureDataLakeHook')
-    @mock.patch('airflow.providers.microsoft.azure.operators.adls_list.AzureDataLakeHook')
+    @mock.patch('airflow.providers.microsoft.azure.operators.adls.AzureDataLakeHook')
     @mock.patch('airflow.providers.google.cloud.transfers.adls_to_gcs.GCSHook')
     def test_execute(self, gcs_mock_hook, adls_one_mock_hook, adls_two_mock_hook):
         """Test the execute function when the run is successful."""
@@ -104,7 +104,7 @@ class TestAdlsToGoogleCloudStorageOperator(unittest.TestCase):
         assert sorted(MOCK_FILES) == sorted(uploaded_files)
 
     @mock.patch('airflow.providers.google.cloud.transfers.adls_to_gcs.AzureDataLakeHook')
-    @mock.patch('airflow.providers.microsoft.azure.operators.adls_list.AzureDataLakeHook')
+    @mock.patch('airflow.providers.microsoft.azure.operators.adls.AzureDataLakeHook')
     @mock.patch('airflow.providers.google.cloud.transfers.adls_to_gcs.GCSHook')
     def test_execute_with_gzip(self, gcs_mock_hook, adls_one_mock_hook, adls_two_mock_hook):
         """Test the execute function when the run is successful."""
